+++ src/tls/lib.rs
@@ -14,18 +14,10 @@
 
 use super::Error;
 
-#[allow(unused_imports)]
-use crate::PQC_ENABLED;
-use crate::TLS12_ENABLED;
 use crate::identity::{self, Identity};
-use crate::strng::Strng;
 
 use std::fmt::Debug;
-
 use std::sync::Arc;
-
-use rustls;
-use rustls::crypto::CryptoProvider;
 
 use rustls::ClientConfig;
 use rustls::ServerConfig;
@@ -40,57 +32,6 @@ pub trait ServerCertProvider: Send + Sync + Clone {
     async fn fetch_cert(&mut self) -> Result<Arc<ServerConfig>, TlsError>;
 }
 
-/// TLS version for runtime configuration.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize)]
-pub enum TlsVersion {
-    Tls12,
-    Tls13,
-}
-
-/// TLS configuration from MeshConfig.meshMTLS.
-#[derive(Debug, Clone, Default, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TlsConfig {
-    pub min_protocol_version: Option<TlsVersion>,
-    pub cipher_suites: Vec<String>,
-    pub ecdh_curves: Vec<String>,
-}
-
-/// Runtime mesh settings from xDS (maps to MeshConfig fields).
-#[derive(Debug, Clone, Default, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MeshSettings {
-    pub trust_domain: Strng,
-    pub trust_domain_aliases: Vec<Strng>,
-    pub tls: Option<TlsConfig>,
-}
-
-/// Check if TLS 1.2 is enabled (xDS settings > env var).
-fn is_tls12_enabled(settings: Option<&MeshSettings>) -> bool {
-    if let Some(ms) = settings {
-        if let Some(ref tls) = ms.tls {
-            if let Some(min_version) = tls.min_protocol_version {
-                return min_version == TlsVersion::Tls12;
-            }
-        }
-    }
-    *TLS12_ENABLED
-}
-
-static TLS_VERSIONS_13_ONLY: &[&rustls::SupportedProtocolVersion] = &[&rustls::version::TLS13];
-static TLS_VERSIONS_12_AND_13: &[&rustls::SupportedProtocolVersion] =
-    &[&rustls::version::TLS13, &rustls::version::TLS12];
-
-pub fn tls_versions(
-    settings: Option<&MeshSettings>,
-) -> &'static [&'static rustls::SupportedProtocolVersion] {
-    if is_tls12_enabled(settings) {
-        TLS_VERSIONS_12_AND_13
-    } else {
-        TLS_VERSIONS_13_ONLY
-    }
-}
-
 #[cfg(feature = "tls-aws-lc")]
 pub static CRYPTO_PROVIDER: &str = "tls-aws-lc";
 #[cfg(feature = "tls-ring")]
@@ -99,260 +40,6 @@ pub static CRYPTO_PROVIDER: &str = "tls-ring";
 pub static CRYPTO_PROVIDER: &str = "tls-boring";
 #[cfg(feature = "tls-openssl")]
 pub static CRYPTO_PROVIDER: &str = "tls-openssl";
-
-// Ztunnel use `rustls` with pluggable crypto modules.
-// All crypto MUST be done via the below providers.
-//
-// One exception is CSR generation which doesn't currently have a plugin mechanism (https://github.com/rustls/rcgen/issues/228);
-// In that case, and any future ones, it is critical to guard the code with appropriate `cfg` guards.
-
-/// Macro to generate cipher suite parsing functions for each crypto provider.
-/// All providers support the same cipher suite names, only the module path differs.
-macro_rules! impl_parse_cipher_suites {
-    ($fn_name:ident, $provider_mod:path) => {
-        fn $fn_name(
-            settings: Option<&MeshSettings>,
-        ) -> Option<Vec<rustls::SupportedCipherSuite>> {
-            use $provider_mod as provider_cs;
-            let tls = settings?.tls.as_ref()?;
-            if tls.cipher_suites.is_empty() {
-                return None;
-            }
-
-            let mut suites = Vec::new();
-            for name in &tls.cipher_suites {
-                match name.as_str() {
-                    "TLS_AES_256_GCM_SHA384" => suites.push(provider_cs::TLS13_AES_256_GCM_SHA384),
-                    "TLS_AES_128_GCM_SHA256" => suites.push(provider_cs::TLS13_AES_128_GCM_SHA256),
-                    "TLS_CHACHA20_POLY1305_SHA256" => {
-                        suites.push(provider_cs::TLS13_CHACHA20_POLY1305_SHA256)
-                    }
-                    "TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384" => {
-                        suites.push(provider_cs::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384)
-                    }
-                    "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256" => {
-                        suites.push(provider_cs::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256)
-                    }
-                    "TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384" => {
-                        suites.push(provider_cs::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384)
-                    }
-                    "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256" => {
-                        suites.push(provider_cs::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256)
-                    }
-                    "TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256" => {
-                        suites.push(provider_cs::TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256)
-                    }
-                    "TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256" => {
-                        suites.push(provider_cs::TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256)
-                    }
-                    unknown => tracing::warn!("Unknown cipher suite '{}', ignoring", unknown),
-                }
-            }
-
-            if suites.is_empty() {
-                tracing::warn!(
-                    "all configured cipher suites were unrecognized ({:?}), falling back to defaults",
-                    tls.cipher_suites
-                );
-                None
-            } else {
-                Some(suites)
-            }
-        }
-    };
-}
-
-#[cfg(feature = "tls-boring")]
-pub(crate) fn provider(_settings: Option<&MeshSettings>) -> Arc<CryptoProvider> {
-    if let Some(ms) = _settings {
-        if ms.tls.is_some() {
-            tracing::warn!("MeshSettings TLS configuration ignored: BoringSSL FIPS uses fixed cipher suites");
-        }
-    }
-    Arc::new(boring_rustls_provider::provider())
-}
-
-#[cfg(feature = "tls-ring")]
-impl_parse_cipher_suites!(parse_cipher_suites_ring, rustls::crypto::ring::cipher_suite);
-
-#[cfg(feature = "tls-ring")]
-pub(crate) fn provider(settings: Option<&MeshSettings>) -> Arc<CryptoProvider> {
-    let cipher_suites = if let Some(suites) = parse_cipher_suites_ring(settings) {
-        suites
-    } else {
-        let mut suites = vec![
-            rustls::crypto::ring::cipher_suite::TLS13_AES_256_GCM_SHA384,
-            rustls::crypto::ring::cipher_suite::TLS13_AES_128_GCM_SHA256,
-        ];
-        if is_tls12_enabled(settings) {
-            suites.extend([
-                rustls::crypto::ring::cipher_suite::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
-                rustls::crypto::ring::cipher_suite::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-                rustls::crypto::ring::cipher_suite::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-                rustls::crypto::ring::cipher_suite::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-            ]);
-        }
-        suites
-    };
-    Arc::new(CryptoProvider {
-        cipher_suites,
-        ..rustls::crypto::ring::default_provider()
-    })
-}
-
-#[cfg(feature = "tls-aws-lc")]
-impl_parse_cipher_suites!(parse_cipher_suites_aws_lc, rustls::crypto::aws_lc_rs::cipher_suite);
-
-#[cfg(feature = "tls-aws-lc")]
-pub(crate) fn provider(settings: Option<&MeshSettings>) -> Arc<CryptoProvider> {
-    let cipher_suites = if let Some(suites) = parse_cipher_suites_aws_lc(settings) {
-        suites
-    } else {
-        let mut suites = vec![
-            rustls::crypto::aws_lc_rs::cipher_suite::TLS13_AES_256_GCM_SHA384,
-            rustls::crypto::aws_lc_rs::cipher_suite::TLS13_AES_128_GCM_SHA256,
-        ];
-        if is_tls12_enabled(settings) {
-            suites.extend([
-                rustls::crypto::aws_lc_rs::cipher_suite::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
-                rustls::crypto::aws_lc_rs::cipher_suite::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-                rustls::crypto::aws_lc_rs::cipher_suite::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-                rustls::crypto::aws_lc_rs::cipher_suite::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-            ]);
-        }
-        suites
-    };
-
-    let mut provider = CryptoProvider {
-        cipher_suites,
-        ..rustls::crypto::aws_lc_rs::default_provider()
-    };
-
-    // Key exchange groups: xDS settings override PQC_ENABLED env var
-    if let Some(kx_groups) = parse_kx_groups_aws_lc(settings) {
-        provider.kx_groups = kx_groups;
-    } else if *PQC_ENABLED {
-        provider.kx_groups = vec![rustls::crypto::aws_lc_rs::kx_group::X25519MLKEM768];
-    }
-
-    Arc::new(provider)
-}
-
-#[cfg(feature = "tls-aws-lc")]
-fn parse_kx_groups_aws_lc(
-    settings: Option<&MeshSettings>,
-) -> Option<Vec<&'static dyn rustls::crypto::SupportedKxGroup>> {
-    let tls = settings?.tls.as_ref()?;
-    if tls.ecdh_curves.is_empty() {
-        return None;
-    }
-
-    let mut groups = Vec::new();
-    for name in &tls.ecdh_curves {
-        match name.as_str() {
-            "X25519MLKEM768" => groups.push(rustls::crypto::aws_lc_rs::kx_group::X25519MLKEM768),
-            "X25519" => groups.push(rustls::crypto::aws_lc_rs::kx_group::X25519),
-            "P-256" | "SECP256R1" => groups.push(rustls::crypto::aws_lc_rs::kx_group::SECP256R1),
-            "P-384" | "SECP384R1" => groups.push(rustls::crypto::aws_lc_rs::kx_group::SECP384R1),
-            unknown => tracing::warn!("Unknown ECDH curve '{}', ignoring", unknown),
-        }
-    }
-
-    if groups.is_empty() {
-        tracing::warn!(
-            "all configured ECDH curves were unrecognized ({:?}), falling back to defaults",
-            tls.ecdh_curves
-        );
-        None
-    } else {
-        Some(groups)
-    }
-}
-
-#[cfg(feature = "tls-openssl")]
-impl_parse_cipher_suites!(parse_cipher_suites_openssl, rustls_openssl::cipher_suite);
-
-#[cfg(feature = "tls-openssl")]
-pub(crate) fn provider(settings: Option<&MeshSettings>) -> Arc<CryptoProvider> {
-    let cipher_suites = if let Some(suites) = parse_cipher_suites_openssl(settings) {
-        suites
-    } else {
-        let mut suites = vec![
-            rustls_openssl::cipher_suite::TLS13_AES_256_GCM_SHA384,
-            rustls_openssl::cipher_suite::TLS13_AES_128_GCM_SHA256,
-        ];
-        if is_tls12_enabled(settings) {
-            suites.extend([
-                rustls_openssl::cipher_suite::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
-                rustls_openssl::cipher_suite::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-                rustls_openssl::cipher_suite::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-                rustls_openssl::cipher_suite::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-            ]);
-        }
-        suites
-    };
-
-    // Key exchange groups: xDS settings > PQC_ENABLED env var > defaults
-    let kx_groups = if let Some(groups) = parse_kx_groups_openssl(settings) {
-        groups
-    } else if *PQC_ENABLED {
-        #[cfg(ossl350)]
-        if openssl::version::number() >= 0x30500000 {
-            vec![rustls_openssl::kx_group::X25519MLKEM768]
-        } else {
-            panic!("COMPLIANCE_POLICY=pqc requires OpenSSL >=3.5.0");
-        }
-        #[cfg(not(ossl350))]
-        panic!("COMPLIANCE_POLICY=pqc requires OpenSSL >=3.5.0");
-    } else {
-        vec![
-            rustls_openssl::kx_group::SECP256R1,
-            rustls_openssl::kx_group::SECP384R1,
-        ]
-    };
-
-    Arc::new(CryptoProvider {
-        cipher_suites,
-        kx_groups,
-        ..rustls_openssl::default_provider()
-    })
-}
-
-#[cfg(feature = "tls-openssl")]
-fn parse_kx_groups_openssl(
-    settings: Option<&MeshSettings>,
-) -> Option<Vec<&'static dyn rustls::crypto::SupportedKxGroup>> {
-    let tls = settings?.tls.as_ref()?;
-    if tls.ecdh_curves.is_empty() {
-        return None;
-    }
-
-    let mut groups = Vec::new();
-    for name in &tls.ecdh_curves {
-        match name.as_str() {
-            #[cfg(ossl350)]
-            "X25519MLKEM768" if openssl::version::number() >= 0x30500000 => {
-                groups.push(rustls_openssl::kx_group::X25519MLKEM768);
-            }
-            "X25519" => groups.push(rustls_openssl::kx_group::X25519),
-            "P-256" | "SECP256R1" => groups.push(rustls_openssl::kx_group::SECP256R1),
-            "P-384" | "SECP384R1" => groups.push(rustls_openssl::kx_group::SECP384R1),
-            unknown => {
-                tracing::warn!("Unknown ECDH curve '{}', ignoring", unknown);
-            }
-        }
-    }
-
-    if groups.is_empty() {
-        tracing::warn!(
-            "all configured ECDH curves were unrecognized ({:?}), falling back to defaults",
-            tls.ecdh_curves
-        );
-        None
-    } else {
-        Some(groups)
-    }
-}
 
 #[derive(thiserror::Error, Debug)]
 pub enum TlsError {
@@ -462,219 +149,5 @@ pub mod tests {
         );
         assert!(!future_certs.is_expired());
         assert_eq!(future_certs.get_duration_until_refresh(), zero_dur);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_openssl_provider_created_successfully() {
-        let provider = super::provider(None);
-        assert!(
-            !provider.kx_groups.is_empty(),
-            "kx_groups should not be empty"
-        );
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_openssl_provider_kx_groups_valid() {
-        let provider = super::provider(None);
-        let expected_len = if *crate::PQC_ENABLED { 1 } else { 2 };
-        assert_eq!(
-            provider.kx_groups.len(),
-            expected_len,
-            "PQC={} should have {} kx groups",
-            *crate::PQC_ENABLED,
-            expected_len
-        );
-    }
-
-    #[test]
-    #[cfg(all(feature = "tls-openssl", not(ossl350)))]
-    fn test_pqc_panic_expected_without_ossl350() {
-        if *crate::PQC_ENABLED {
-            panic!("PQC_ENABLED=true without ossl350 cfg - provider() will panic");
-        }
-    }
-
-    #[test]
-    fn test_mesh_settings_default() {
-        let settings = super::MeshSettings::default();
-        assert!(settings.trust_domain.is_empty());
-        assert!(settings.trust_domain_aliases.is_empty());
-        assert!(settings.tls.is_none());
-    }
-
-    #[test]
-    fn test_is_tls12_enabled_none_settings() {
-        let result = super::is_tls12_enabled(None);
-        assert_eq!(result, *crate::TLS12_ENABLED);
-    }
-
-    #[test]
-    fn test_is_tls12_enabled_with_tls13() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec![],
-                ecdh_curves: vec![],
-            }),
-        };
-        assert!(!super::is_tls12_enabled(Some(&settings)));
-    }
-
-    #[test]
-    fn test_is_tls12_enabled_with_tls12() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls12),
-                cipher_suites: vec![],
-                ecdh_curves: vec![],
-            }),
-        };
-        assert!(super::is_tls12_enabled(Some(&settings)));
-    }
-
-    #[test]
-    fn test_is_tls12_enabled_no_tls_config() {
-        let settings = super::MeshSettings {
-            trust_domain: "cluster.local".into(),
-            trust_domain_aliases: vec![],
-            tls: None,
-        };
-        assert_eq!(super::is_tls12_enabled(Some(&settings)), *crate::TLS12_ENABLED);
-    }
-
-    #[test]
-    fn test_tls_versions_with_tls13() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec![],
-                ecdh_curves: vec![],
-            }),
-        };
-        let versions = super::tls_versions(Some(&settings));
-        assert_eq!(versions.len(), 1);
-        assert_eq!(versions[0], &rustls::version::TLS13);
-    }
-
-    #[test]
-    fn test_tls_versions_with_tls12() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls12),
-                cipher_suites: vec![],
-                ecdh_curves: vec![],
-            }),
-        };
-        let versions = super::tls_versions(Some(&settings));
-        assert_eq!(versions.len(), 2);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_provider_with_mesh_settings_tls13() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec![],
-                ecdh_curves: vec![],
-            }),
-        };
-        let provider = super::provider(Some(&settings));
-        assert_eq!(provider.cipher_suites.len(), 2);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_provider_with_mesh_settings_tls12() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls12),
-                cipher_suites: vec![],
-                ecdh_curves: vec![],
-            }),
-        };
-        let provider = super::provider(Some(&settings));
-        assert_eq!(provider.cipher_suites.len(), 6);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_provider_with_custom_cipher_suites() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec!["TLS_AES_256_GCM_SHA384".to_string()],
-                ecdh_curves: vec![],
-            }),
-        };
-        let provider = super::provider(Some(&settings));
-        assert_eq!(provider.cipher_suites.len(), 1);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_provider_with_custom_ecdh_curves() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec![],
-                ecdh_curves: vec!["P-256".to_string(), "P-384".to_string()],
-            }),
-        };
-        let provider = super::provider(Some(&settings));
-        assert_eq!(provider.kx_groups.len(), 2);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_provider_ignores_unknown_cipher_suites() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec![
-                    "UNKNOWN_CIPHER".to_string(),
-                    "TLS_AES_256_GCM_SHA384".to_string(),
-                ],
-                ecdh_curves: vec![],
-            }),
-        };
-        let provider = super::provider(Some(&settings));
-        assert_eq!(provider.cipher_suites.len(), 1);
-    }
-
-    #[test]
-    #[cfg(feature = "tls-openssl")]
-    fn test_provider_all_unknown_cipher_suites_falls_back() {
-        let settings = super::MeshSettings {
-            trust_domain: Default::default(),
-            trust_domain_aliases: vec![],
-            tls: Some(super::TlsConfig {
-                min_protocol_version: Some(super::TlsVersion::Tls13),
-                cipher_suites: vec!["UNKNOWN_CIPHER".to_string()],
-                ecdh_curves: vec![],
-            }),
-        };
-        let provider = super::provider(Some(&settings));
-        assert_eq!(provider.cipher_suites.len(), 2);
     }
 }
